+++ src/dashboard/files/ExamplesRoute.tsx
@@ -1,3 +1,4 @@
+import { Divider } from '@mui/material';
 import { EXAMPLE_FILES } from '../../constants/appConstants';
 import { ROUTES } from '../../constants/routes';
 import { DashboardFileLink } from '../components/DashboardFileLink';
@@ -8,7 +9,15 @@ export const Component = () => {
     <>
       <DashboardHeader title="Examples" />
       {Object.entries(EXAMPLE_FILES).map(([id, { name, description }]) => (
-        <DashboardFileLink key={id} name={name} description={description} to={`${ROUTES.CREATE_FILE}?example=${id}`} />
+        <>
+          <DashboardFileLink
+            key={id}
+            name={name}
+            description={description}
+            to={`${ROUTES.CREATE_FILE}?example=${id}`}
+          />
+          <Divider />
+        </>
       ))}
     </>
   );
