+++ src/debugFlags.ts
@@ -65,3 +65,9 @@ export const debugShowCellsForDirtyQuadrants = debug && false;
 export const debugShowFileIO = debug && false;
 
 export const debugGridSettings = debug && false;
+
+// --------
+// UI
+// --------
+
+export const debugShowUILogs = debug && false;
